pyproject.toml
[tool.poetry]
name = "kafka-task"
version = "0.1.0"
description = "Example of kafka usage via simulation of IoT workers"
authors = ["Victoria <victoria2000dzuba@gmail.com>"]
readme = "README.md"
packages = [{include = "kafka_task"}]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
pyyaml = "^6.0.1"
numpy = "^1.26.2"
pandas = "^2.1.4"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
isort = "^5.13.2"
black = "^23.12.0"

[tool.poetry.scripts]
producers = "kafka_task.producer:main"
consumers = "kafka_task.consumer:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
